refactor: detach view logic from domain logic
- view 로직을 domain 로직에서 분리하여 개발하였습니다.
- 소스들을 용도에 맞는 패키지로 분리하여 넣었습니다.
