+++ README.md
@@ -1,5 +1,12 @@
 # 자동차 경주 게임
 
+# Step5
+## 요구사항
+* MVC 패턴 기반으로 리팩토링해 view 패키지의 객체가 domain 패키지 객체에 의존할 수 있지만, domain 패키지의 객체는 view 패키지 객체에 의존하지 않도록 구현한다.
+
+## 진행 목록
+-[x] 도메인 로직에 있는 view 로직을 찾아 리팩토링 한다.
+
 # Step4
 ## 기능 요구사항
 * 각 자동차에 이름을 부여할 수 있다. 자동차 이름은 5자를 초과할 수 없다.
