--- src/main/java/step4/ConsoleOutputView.java
@@ -1,8 +1,0 @@
-package step4;
-
-public class ConsoleOutputView implements OutputView {
-    @Override
-    public void print(String message) {
-        System.out.println(message);
-    }
-}
