--- src/main/java/step4/OutputView.java
@@ -1,5 +1,0 @@
-package step4;
-
-public interface OutputView {
-     void print(String message);
-}
