+++ src/main/java/step4/domain/Car.java
@@ -1,4 +1,4 @@
-package step4;
+package step4.domain;
 
 public class Car {
     private final String carName;
