+++ src/main/java/step4/domain/RacingCarFactory.java
@@ -1,11 +1,13 @@
-package step4;
+package step4.domain;
+
+import step4.view.RacingCarInputView;
 
 import java.util.Arrays;
 import java.util.List;
 
 public class RacingCarFactory {
 
-    static RacingCar createRacingCarWithInputView() {
+    public static RacingCar createRacingCarWithInputView() {
 
         String carInfo = RacingCarInputView.scanCarInfo();
         int numberOfRace = RacingCarInputView.scanNumberOfRace();
