+++ src/main/java/step4/view/OutputView.java
@@ -1,0 +1,8 @@
+package step4.view;
+
+import step4.domain.RacingCar;
+
+public interface OutputView {
+     void print(String message);
+     void startAndPrintRacing(RacingCar racingCar);
+}
